requirements.txt
black
requests
spotipy
yt-dlp
python-dateutil
selenium
beautifulsoup4
ytmusicapi